fixing routin issues
