+++ file-sharing-app/app/(dashboard)/(routes)/files/_components/FilePreview.js
