+++ file-sharing-app/app/(dashboard)/(routes)/files/_components/ProgressBar.js
