+++ file-sharing-app/next.config.mjs
@@ -1,6 +1,8 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  
+    images: {
+        domains: ['firebasestorage.googleapis.com']
+    }
 };
 
 export default nextConfig;
